Move join-request / accept tests to loraserver_test.
